+++ packages/sdks/web-component/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.44.3](https://github.com/descope/descope-js/compare/web-component-3.44.2...web-component-3.44.3) (2025-07-31)
+
+
+### Bug Fixes
+
+* add logs to redirect in popup ([#1174](https://github.com/descope/descope-js/issues/1174)) ([f8b8eb1](https://github.com/descope/descope-js/commit/f8b8eb1b8ebb4ba56b53dd3e3fbf9505994776dd))
+
+## [3.44.2](https://github.com/descope/descope-js/compare/web-component-3.44.1...web-component-3.44.2) (2025-07-31)
+
+
+### Bug Fixes
+
+* trying to fix issue 11567 RELEASE ([#1173](https://github.com/descope/descope-js/issues/1173)) ([6515f53](https://github.com/descope/descope-js/commit/6515f53abf5a5d51cb7607bda136cdbb19dc349e))
+
 ## [3.44.1](https://github.com/descope/descope-js/compare/web-component-3.44.0...web-component-3.44.1) (2025-07-27)